// src/app/Model/Contributeur.ts
export class Contributeur{
  idC: number;
  nameC: string;
  emailC: string;
  birthDayC: string;
  nationalityC: string;
  occupationC: string;
  phoneNumberC: number;
  decision: string;
  linkBehance: string;
  linkDribble : string;
  linkLinkedin : string;
  linkFacebook : string;
  filePortfolio : string;

}